moved schemas to modules folder + exported
