+++ server/models/Cohort.js
@@ -1,0 +1,52 @@
+const mongoose = require("mongoose");
+const Schema = mongoose.Schema;
+
+const cohortsSchema = new Schema({
+    cohortSlug: {
+      type: String,
+      required: true,
+    },
+    cohortName: {
+      type: String,
+      required: true,
+    },
+    program: {
+      type: String,
+      enum: ["Web Dev", "UX/UI", "Data Analytics", "Cybersecurity"],
+    },
+    format: {
+      type: String,
+      enum: ["Full Time", "Part Time"],
+    },
+    campus:{
+      type: String,
+      enum: ["Madrid", "Barcelona", "Miami", "Paris", "Berlin", "Amsterdam", "Lisbon", "Remote"],
+    },
+    startDate: {
+      type: Date,
+      default: Date.now(),
+    },
+    endDate: {
+      type: Date,
+    },
+    inProgress:{
+      type: Boolean,
+      default: false,
+    },
+    programManager:{
+      type: String,
+      required: true,
+    },
+    leadTeacher:{
+      type: String,
+      required: true,
+    },
+    totalHours:{
+      type: Number,
+      default: 360,
+    },
+  });
+
+  const Cohort = mongoose.model("Cohort", cohortsSchema);
+
+module.exports = Cohort
